Add files via upload
Included Search-and-replace functionality for all "input()", as this made it impossible to remotely execute code that contained input() functions. For code that somehow contained the prefix for my regex pattern "input(", but was in fact not an input() that needed to be replaced - you can bypass this and execute your code as is with the command "!c! false positive". Also, included try/except for code execution, errors will not appear in console upon exec() if code is not valid.